+++ backend/substrapp/tests/views/tests_views_algo.py
@@ -5,7 +5,6 @@ import logging
 import json
 
 import mock
-import unittest
 from parameterized import parameterized
 
 from django.urls import reverse
@@ -19,7 +18,7 @@ from orchestrator.client import OrchestratorClient
 from grpc import RpcError, StatusCode
 
 from ..common import get_sample_algo, AuthenticatedClient, encode_filter
-from ..assets import objective, datamanager, algo, model
+from ..assets import algo
 
 MEDIA_ROOT = "/tmp/unittests_views/"
 
@@ -81,31 +80,6 @@ class AlgoViewTests(APITestCase):
         with mock.patch.object(OrchestratorClient, 'query_algos', return_value=algo):
             search_params = f'?search=algo%253Aname%253A{encode_filter(algo[2]["name"])}'
             search_params += f'%2Calgo%253Aowner%253A{encode_filter(algo[2]["owner"])}'
-            response = self.client.get(self.url + search_params, **self.extra)
-            r = response.json()
-
-            self.assertEqual(len(r['results']), 1)
-
-    def test_algo_list_filter_datamanager_fail(self):
-        with mock.patch.object(OrchestratorClient, 'query_algos', return_value=algo), \
-                mock.patch.object(OrchestratorClient, 'query_datamanagers', return_value=datamanager):
-            search_params = '?search=dataset%253Aname%253ASimplified%2520ISIC%25202018'
-            response = self.client.get(self.url + search_params, **self.extra)
-            self.assertIn('Malformed search filters', response.json()['message'])
-
-    def test_algo_list_filter_objective_fail(self):
-        with mock.patch.object(OrchestratorClient, 'query_algos', return_value=algo), \
-                mock.patch.object(OrchestratorClient, 'query_objectives', return_value=objective):
-            search_params = '?search=objective%253Aname%253ASkin%2520Lesion%2520Classification%2520Objective'
-            response = self.client.get(self.url + search_params, **self.extra)
-            self.assertIn('Malformed search filters', response.json()['message'])
-
-    @unittest.skip("filter on model key does not work anymore")
-    def test_algo_list_filter_model(self):
-
-        with mock.patch.object(OrchestratorClient, 'query_algos', return_value=algo), \
-                mock.patch.object(OrchestratorClient, 'query_models', return_value=model):
-            search_params = f'?search=model%253Akey%253A{model[0]["key"]}'
             response = self.client.get(self.url + search_params, **self.extra)
             r = response.json()