+++ backend/substrapp/views/algo.py
@@ -198,7 +198,6 @@ class AlgoViewSet(mixins.CreateModelMixin,
         if query_params is not None:
             try:
                 data = filter_list(
-                    channel_name=get_channel_name(request),
                     object_type='algo',
                     data=data,
                     query_params=query_params)