+++ backend/substrapp/views/computetask.py
@@ -170,7 +170,6 @@ class ComputeTaskViewSet(mixins.CreateModelMixin,
         if query_params is not None:
             try:
                 data = filter_list(
-                    channel_name=get_channel_name(request),
                     object_type=self.basename,
                     data=data,
                     query_params=query_params)