+++ backend/substrapp/views/datamanager.py
@@ -207,7 +207,6 @@ class DataManagerViewSet(mixins.CreateModelMixin,
         if query_params is not None:
             try:
                 data = filter_list(
-                    channel_name=get_channel_name(request),
                     object_type='dataset',
                     data=data,
                     query_params=query_params)