+++ backend/substrapp/views/model.py
@@ -98,7 +98,6 @@ class ModelViewSet(PaginationMixin,
         if query_params is not None:
             try:
                 data = filter_list(
-                    channel_name=get_channel_name(request),
                     object_type="model",
                     data=data,
                     query_params=query_params,