+++ backend/substrapp/views/objective.py
@@ -193,7 +193,6 @@ class ObjectiveViewSet(mixins.CreateModelMixin,
         if query_params is not None:
             try:
                 data = filter_list(
-                    channel_name=get_channel_name(request),
                     object_type='objective',
                     data=data,
                     query_params=query_params)